initial 64 bit update (incomplete)

+++ kernel/io/idt.h
@@ -1,7 +1,9 @@
 #ifndef IDT_H
 #define IDT_H
 
+#include <stdint.h>
+
 void idt_init();
-void idt_set_entry(int n, unsigned int handler);
+void idt_set_entry(int n, uint64_t handler);
 
 #endif
